+++ src/main/java/com/electronicstore/service/SupplierService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 @Service
-public class SupplierService  implements ISupplierService {
+public class SupplierService  extends BaseService implements ISupplierService {
     @Autowired
     private SupplierRepository supplierRepository;
     @Transactional
@@ -21,10 +21,16 @@ public class SupplierService  implements ISupplierService {
 
     @Transactional
     public Supplier addSupplier(Supplier supplier) {
-        return supplierRepository.save(supplier);
+        if(supplier==null|| supplier.getName() == null || supplier.getName().trim().isEmpty()) {
+            throw new IllegalArgumentException("Supplier name cannot be empty");
+        }
+        User user = getAuthenticatedUser();
+        Supplier savedSupplier = supplierRepository.save(supplier);
+        return savedSupplier;
     }
     @Transactional
     public Supplier updateSupplier(Supplier supplier) {
+        User user = getAuthenticatedUser();
         Optional<Supplier> supplierOptional= supplierRepository.findById(supplier.getId());
    if (!supplierOptional.isPresent()) {
            throw new IllegalArgumentException("produkti me ID" + supplier.getId() + "nuk gjendet");
@@ -38,6 +44,17 @@ public class SupplierService  implements ISupplierService {
     }
     @Transactional
     public void deleteSupplier(long id) {
+        User user = getAuthenticatedUser();
+        Optional<Supplier> supplierOptional= supplierRepository.findById(id);
+        if (supplierOptional.isEmpty()) {
+            throw new IllegalArgumentException("Supplier not found with id: " + id);
+        }
+
+        Supplier supplier = supplierOptional.get();
+
+        supplier.getItems().forEach(item -> item.getSuppliers().remove(supplier));
+        supplier.getItems().clear();
+
 
         supplierRepository.deleteById(id);
     }
